add slugs to posts, and now only admin or staff can create,edit,delete the posts

+++ posts/views.py
@@ -1,12 +1,14 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from django.contrib import messages
 # from django.http import HttpResponse
 from .models import Post
 from .forms import PostForm
 # Create your views here.
 def post_create(request):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -19,8 +21,8 @@ def post_create(request):
     }
     return render(request, "post_form.html", context)
 
-def post_details(request, id):
-    instance = get_object_or_404(Post, id=id)
+def post_details(request, slug):
+    instance = get_object_or_404(Post, slug=slug)
     context = {
         "title": instance.title,
         "instance": instance
@@ -48,8 +50,10 @@ def post_list(request):
     }
     return render(request, 'post_list.html', context)
 
-def post_update(request, id):
-    instance = get_object_or_404(Post, id=id)
+def post_update(request, slug):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None,request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -62,8 +66,10 @@ def post_update(request, id):
     }
     return render(request, "post_form.html", context)
 
-def post_delete(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Post deleted")
     return redirect("posts:list")
